APS :: Adicionado sistema checkpoint

+++ Services/ILocalService.cs
@@ -5,6 +5,6 @@ namespace GuieMe.Services
     public interface ILocalService
     {
         List<Local> Locais();
-        List<Tuple<decimal, decimal>> Paredes();
+        List<Tuple<int, int>> Paredes();
     }
 }
